requirements.txt
python-socketio == 4.6.0
gevent == 20.9.0